app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
alembic
passlib[bcrypt]
python-jose[cryptography]
python-multipart